Refactor dynamic shape analysis
